Extract GeoTiff generation
