+++ src/ext/java/org/opentripplanner/ext/traveltime/TravelTimeResource.java
@@ -1,7 +1,5 @@
 package org.opentripplanner.ext.traveltime;
 
-import static javax.imageio.ImageWriteParam.MODE_EXPLICIT;
-
 import jakarta.ws.rs.DefaultValue;
 import jakarta.ws.rs.GET;
 import jakarta.ws.rs.Path;
@@ -11,7 +9,6 @@ import jakarta.ws.rs.core.Context;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
 import jakarta.ws.rs.core.StreamingOutput;
-import java.awt.image.DataBuffer;
 import java.time.Instant;
 import java.time.LocalDate;
 import java.time.Period;
@@ -22,22 +19,7 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import java.util.Set;
-import javax.media.jai.RasterFactory;
-import org.geotools.coverage.grid.GridCoverage2D;
-import org.geotools.coverage.grid.GridCoverageFactory;
-import org.geotools.coverage.grid.GridEnvelope2D;
-import org.geotools.coverage.grid.GridGeometry2D;
-import org.geotools.coverage.grid.io.AbstractGridFormat;
 import org.geotools.data.geojson.GeoJSONWriter;
-import org.geotools.gce.geotiff.GeoTiffFormat;
-import org.geotools.gce.geotiff.GeoTiffWriteParams;
-import org.geotools.gce.geotiff.GeoTiffWriter;
-import org.geotools.geometry.Envelope2D;
-import org.geotools.referencing.crs.DefaultGeographicCRS;
-import org.geotools.referencing.operation.transform.AffineTransform2D;
-import org.locationtech.jts.geom.Coordinate;
-import org.opengis.parameter.GeneralParameterValue;
-import org.opengis.parameter.ParameterValueGroup;
 import org.opentripplanner.api.common.LocationStringParser;
 import org.opentripplanner.api.parameter.QualifiedModeSet;
 import org.opentripplanner.ext.traveltime.geometry.ZSampleGrid;
@@ -174,52 +156,7 @@ public class TravelTimeResource {
   @Produces("image/tiff")
   public Response getSurface() {
     ZSampleGrid<WTWD> sampleGrid = getSampleGrid();
-
-    int minX = sampleGrid.getXMin();
-    int minY = sampleGrid.getYMin();
-    int maxY = sampleGrid.getYMax();
-
-    int width = sampleGrid.getXMax() - minX + 1;
-    int height = maxY - minY + 1;
-
-    Coordinate center = sampleGrid.getCenter();
-
-    double resX = sampleGrid.getCellSize().x;
-    double resY = sampleGrid.getCellSize().y;
-
-    var raster = RasterFactory.createBandedRaster(DataBuffer.TYPE_INT, width, height, 1, null);
-    var dataBuffer = raster.getDataBuffer();
-
-    // Initialize with NO DATA value
-    for (int i = 0; i < dataBuffer.getSize(); i++) {
-      dataBuffer.setElem(i, Integer.MIN_VALUE);
-    }
-
-    for (var s : sampleGrid) {
-      final WTWD z = s.getZ();
-      raster.setSample(s.getX() - minX, maxY - s.getY(), 0, z.wTime / z.w);
-    }
-
-    Envelope2D geom = new GridGeometry2D(
-      new GridEnvelope2D(0, 0, width, height),
-      new AffineTransform2D(resX, 0, 0, resY, center.x + resX * minX, center.y + resY * minY),
-      DefaultGeographicCRS.WGS84
-    )
-      .getEnvelope2D();
-
-    GridCoverage2D gridCoverage = new GridCoverageFactory().create("traveltime", raster, geom);
-
-    GeoTiffWriteParams wp = new GeoTiffWriteParams();
-    wp.setCompressionMode(MODE_EXPLICIT);
-    wp.setCompressionType("LZW");
-    ParameterValueGroup params = new GeoTiffFormat().getWriteParameters();
-    params.parameter(AbstractGridFormat.GEOTOOLS_WRITE_PARAMS.getName().toString()).setValue(wp);
-    StreamingOutput streamingOutput = outputStream -> {
-      GeoTiffWriter writer = new GeoTiffWriter(outputStream);
-      writer.write(gridCoverage, params.values().toArray(new GeneralParameterValue[1]));
-      writer.dispose();
-      outputStream.close();
-    };
+    StreamingOutput streamingOutput = RasterRenderer.createGeoTiffRaster(sampleGrid);
     return Response.ok().entity(streamingOutput).build();
   }
